+++ src/user.jl
@@ -3,5 +3,6 @@ usermatrixclass =
 
 
 
+"mysymmtric" => ["hilb", "cauchy", "circul", "dingdong", "invhilb", "moler", "pascal", "pei", "clement", "fiedler", "minij", "tridiag", "lehmer", "randcorr", "poisson", "wilkinson", "kms", "wathen", "oscillate", "prolate", "deriv2", "shaw", ],
 );